requirements.txt
pyyaml
nnmnkwii
scipy
torch
torchvision
torchaudio
